Add sparse to symbol namespace
