+++ src/operator/tensor/sparse_retain.cu
@@ -26,7 +26,7 @@
 namespace mxnet {
 namespace op {
 
-NNVM_REGISTER_OP(sparse_retain)
+NNVM_REGISTER_OP(_sparse_retain)
 .set_attr<FComputeEx>("FComputeEx<gpu>", SparseRetainOpForwardEx<gpu>);
 
 NNVM_REGISTER_OP(_backward_sparse_retain)
